♻️ refactor : 무한 스크롤에 사용되는 observer 함수 분리
